dataset/architectural_views_rep-pkg/architectural_views_rep-pkg/src/requirements.txt
matplotlib==3.8.1
pandas==2.1.1
PyGithub==1.59.1
tabulate==0.9.0
urllib3==2.0.4
